+++ src/main/java/net/kccricket/bestesttool/security/Permissions.java
@@ -17,6 +17,7 @@ public final class Permissions {
     public static final String PERM_REFILL = "bestesttool.refill";
     public static final String PERM_RELOAD = "bestesttool.reload";
     public static final String PERM_DEBUG = "bestesttool.debug";
+    public static final String PERM_SELFTEST = "bestesttool.selftest";
 
     /**
      * Check if the sender has the specified permission node.